+++ src/examples/chapter8_refactor_test_debug/chainOfReponsibility824/HeaderTextProcessing.java
@@ -1,4 +1,4 @@
-package examples.chainOfReponsibility824;
+package examples.chapter8_refactor_test_debug.chainOfReponsibility824;
 
 public class HeaderTextProcessing extends ProcessingObject<String> {
