+++ src/examples/chapter8_refactor_test_debug/factory825/Loan.java
@@ -1,0 +1,4 @@
+package examples.chapter8_refactor_test_debug.factory825;
+
+public class Loan extends Product {
+}
