+++ src/examples/chapter8_refactor_test_debug/factory825/Stock.java
@@ -1,0 +1,4 @@
+package examples.chapter8_refactor_test_debug.factory825;
+
+public class Stock extends Product {
+}
