+++ src/examples/chapter8_refactor_test_debug/lambdaTesting830/Point.java
@@ -1,4 +1,4 @@
-package examples.lambdaTesting830;
+package examples.chapter8_refactor_test_debug.lambdaTesting830;
 
 import java.util.Comparator;
