--- src/examples/factory825/Bond.java
@@ -1,4 +1,0 @@
-package examples.factory825;
-
-public class Bond extends Product {
-}
