--- src/examples/factory825/Loan.java
@@ -1,4 +1,0 @@
-package examples.factory825;
-
-public class Loan extends Product {
-}
