--- src/examples/factory825/Product.java
@@ -1,4 +1,0 @@
-package examples.factory825;
-
-public class Product {
-}
