--- src/examples/factory825/Stock.java
@@ -1,4 +1,0 @@
-package examples.factory825;
-
-public class Stock extends Product {
-}
